Launch phantomjs using spawn; should work better on Linux/Docker
We have to sanitize /dev/stdin by using the cat | trick
Gory details:  https://groups.google.com/forum/#!msg/nodejs/SxNKLclbM5k/IrgMQ3UDIQAJ

+++ tools/selftest.js
@@ -849,7 +849,8 @@ _.extend(SimpleProcess.prototype, {
     self.process.on('close', function (exitCode) {
       self.exitCode = exitCode;
 
-      // XXX: Disable through an option
+      // XXX: Disable this log message through an option
+      // (though it is a useful default!)
       if (exitCode != 0) {
         console.log("Unexpected exit code", exitCode, "from", self.command, self.args, "\nstdout:\n", self.stdout, "\nstderr:\n", self.stderr);
       }
@@ -874,7 +875,7 @@ _.extend(SimpleProcess.prototype, {
 
   kill: function () {
     var self = this;
-    self.process.kill();
+    if (self.process) self.process.kill();
   }
 });
 
@@ -894,12 +895,13 @@ _.extend(PhantomClient.prototype, {
     var self = this;
 
     var phantomScript = "require('webpage').create().open('" + self.url + "');";
-    var phantomPath = phantomjs.path;
-    self.process = child_process.execFile(
-      '/bin/bash',
-      ['-c',
-       ("exec " + phantomPath + " --load-images=no /dev/stdin <<'END'\n" +
-        phantomScript + "\nEND\n")]);
+
+    // We need to 'cat | ' to sanitize stdin: https://groups.google.com/forum/#!msg/nodejs/SxNKLclbM5k/IrgMQ3UDIQAJ
+    var command = 'cat | ' + phantomjs.path + ' --load-images=no /dev/stdin';
+    self.process = new SimpleProcess('/bin/bash', ['-c', command ]);
+    self.process.start();
+    self.process.stdin.write(phantomScript + "\n");
+    self.process.stdin.end();
   },
 
   stop: function() {
